test(config): split config tests into files
