+++ packages/svrx/__tests__/spec/configure/cli.js
@@ -1,23 +1,5 @@
 const expect = require('expect.js');
-const libPath = require('path');
-const sinon = require('sinon');
-const Option = require('../../lib/configure/option');
-const CONFIGS = require('../../lib/config-list');
-const { BUILTIN_PLUGIN } = require('../../lib/constant');
-const defaults = require('../../lib/util/jsonSchemaDefaults');
-const { createServer } = require('../util');
-
-const TEST_PLUGIN_PATH = libPath.join(__dirname, '../fixture/plugin/svrx-plugin-test');
-const BUILTIN_DEFAULTS = defaults({
-  type: 'object',
-  properties: CONFIGS,
-});
-
-function requireEnsure(path) {
-  delete require.cache[path];
-  /* eslint-disable global-require, import/no-dynamic-require */
-  return require(path);
-}
+const { createServer } = require('../../util');
 
 describe('CLI Config', () => {
   it('should replace alias with real param name', () => {
@@ -300,255 +282,5 @@ describe('CLI Config', () => {
       expect(testPlugin).not.to.be(undefined);
       expect(testPlugin2).not.to.be(undefined);
     });
-  });
-});
-
-describe('Inline/RC File Config', () => {
-  describe('plugins', () => {
-    it('should pick string plugins correctly', () => {
-      const server = createServer({
-        plugins: [
-          'test', 'demo',
-        ],
-      });
-      const testPlugin = server.config.getPlugin('test');
-      const demoPlugin = server.config.getPlugin('demo');
-      expect(testPlugin).not.to.be(undefined);
-      expect(demoPlugin).not.to.be(undefined);
-    });
-
-    it('should pick obj plugins correctly', () => {
-      const server = createServer({
-        plugins: [
-          {
-            name: 'test',
-          },
-        ],
-      });
-      const testPlugin = server.config.getPlugin('test');
-      expect(testPlugin).not.to.be(undefined);
-    });
-
-    it('should parse local plugin name correctly', () => {
-      const server = createServer({
-        plugins: [
-          {
-            path: TEST_PLUGIN_PATH,
-          },
-        ],
-      });
-      const testPlugin = server.config.getPlugin('test');
-      expect(testPlugin).not.to.be(undefined);
-    });
-  });
-});
-
-describe('Builtin Configs', () => {
-  it('should fill all default values', () => {
-    const server = createServer();
-    Object.keys(CONFIGS).forEach((key) => {
-      const value = CONFIGS[key];
-      if (value.default !== undefined && key !== 'open' && key
-        !== 'livereload') {
-        expect(server.config.get(key)).to.eql(value.default);
-      }
-    });
-  });
-
-  it('should enable all builtin plugins by default', () => {
-    const server = createServer();
-    const plugins = server.config.getPlugins();
-    expect(plugins.length).to.eql(BUILTIN_PLUGIN.length);
-    expect(plugins.map((p) => p.getInfo('name'))).to.eql(BUILTIN_PLUGIN);
-  });
-
-  it('should concat array values from CLI and RC', () => {
-    const server = createServer({
-      proxy: ['a'],
-    }, {
-      proxy: ['b'],
-    });
-    expect(server.config.get('proxy')).to.eql(['a', 'b']);
-  });
-});
-
-describe('Config get', () => {
-  const server = createServer({
-    port: 3000,
-    plugins: [
-      {
-        name: 'test',
-        version: '0.0.1',
-        inplace: true,
-        configSchema: {
-          foo: {
-            type: 'string',
-            default: 'bar',
-          },
-        },
-        options: {
-          op: 123,
-        },
-      },
-    ],
-  });
-  const { config } = server;
-  const testPlugin = config.getPlugin('test');
-
-  it('should get builtin value corrently', () => {
-    expect(config.get('port')).to.equal(3000);
-  });
-
-  it('should get plugin info corrently', () => {
-    expect(testPlugin.getInfo('version')).to.equal('0.0.1');
-  });
-
-  it('should get plugin option corrently', () => {
-    expect(testPlugin.get('op')).to.equal(123);
-  });
-
-  it('should get builtin options in plugin config with $', () => {
-    expect(testPlugin.get('$.port')).to.equal(3000);
-  });
-
-  it('should get all options (includes the defaults) when there\'s no path', () => {
-    expect(config.get()).to.eql({
-      ...BUILTIN_DEFAULTS, port: 3000, open: false, livereload: false,
-    });
-  });
-
-  it('should get all plugin options (includes the defaults) when there\'s no path', async () => {
-    await server.setup();
-    expect(testPlugin.get()).to.eql({
-      op: 123,
-      foo: 'bar', // defaults
-    });
-  });
-});
-
-describe('Config set', () => {
-  const server = createServer({
-    port: 3000,
-    plugins: [
-      {
-        name: 'test',
-        version: '0.0.1',
-        inplace: true,
-        configSchema: {
-          foo: {
-            type: 'string',
-            default: 'bar',
-          },
-        },
-        options: {
-          op: 123,
-        },
-      },
-    ],
-  });
-  const { config } = server;
-  const testPlugin = config.getPlugin('test');
-
-  it('should set builtin value corrently', () => {
-    config.set('port', 4000);
-    expect(config.get('port')).to.equal(4000);
-    config.set('port', 3000);
-  });
-
-  it('should set plugin option corrently', () => {
-    testPlugin.set('op', 321);
-    testPlugin.set('other', 'other info');
-    expect(testPlugin.get('op')).to.equal(321);
-    expect(testPlugin.get('other')).to.equal('other info');
-  });
-});
-
-describe('Config Validate', () => {
-  it('should validate single type configs', () => {
-    const option = new Option({
-      https: 3000, // should be boolean
-      port: 'port', // should be number
-      svrx: 123, // should be string
-      urls: 'should be object', // should be object
-    });
-    const errors = option._validate(CONFIGS);
-    expect(errors).not.to.equal(null);
-    expect(errors.length).to.equal(4);
-    expect(errors[0]).to.equal('Config Error: .https should be boolean');
-    expect(errors[1]).to.equal('Config Error: .port should be number');
-    expect(errors[2]).to.equal('Config Error: .svrx should be string');
-    expect(errors[3]).to.equal('Config Error: .urls should be object');
-  });
-
-  it('should validate multi type configs', () => {
-    const option = new Option({
-      proxy: 123, // boolean,array,object
-      open: ['a', 'b'], // boolean,string
-      serve: 'string',
-    });
-    const errors = option._validate(CONFIGS);
-    expect(errors).not.to.equal(null);
-    expect(errors.length).to.equal(3);
-    expect(errors[0])
-      .to
-      .equal('Config Error: .open should be boolean or string');
-    expect(errors[1])
-      .to
-      .equal('Config Error: .proxy should be boolean or object or array');
-    expect(errors[2])
-      .to
-      .equal('Config Error: .serve should be boolean or object');
-  });
-
-  it('should log the error path correctly', () => {
-    const option = new Option({
-      serve: {
-        base: 123, // string
-      },
-      livereload: {
-        exclude: true,
-      },
-    });
-    const errors = option._validate(CONFIGS);
-    expect(errors).not.to.equal(null);
-    expect(errors.length).to.equal(2);
-    expect(errors[0])
-      .to
-      .equal('Config Error: .livereload.exclude should be string or array');
-    expect(errors[1]).to.equal('Config Error: .serve.base should be string');
-  });
-});
-
-describe('Plugin Config', () => {
-  it('should return default values when get plugin(load from path) option', async () => {
-    const server = createServer({
-      plugins: [{
-        path: TEST_PLUGIN_PATH,
-      }],
-    });
-    await server.setup();
-    const testPlugin = server.config.getPlugin('test');
-    expect(testPlugin.get('limit')).to.equal(100);
-  });
-
-  it('should return default values when get plugin(load from remote) option', async () => {
-    const server = createServer({
-      plugins: ['remote'],
-    });
-    const fakeLoadOne = sinon.fake.resolves({
-      name: 'remote',
-      path: TEST_PLUGIN_PATH,
-      module: requireEnsure(TEST_PLUGIN_PATH),
-      version: requireEnsure(TEST_PLUGIN_PATH).version,
-      pluginConfig: server.config.getPlugin('remote'),
-    });
-
-    const pluginDetail = await fakeLoadOne();
-    await server.system.buildOne(pluginDetail);
-
-    const testPlugin = server.config.getPlugin('remote');
-    expect(testPlugin.get('limit')).to.equal(100);
-    expect(testPlugin.get('$.port')).to.equal(8000);
-    sinon.restore();
   });
 });
